// Iserv.Niis.Angular/src/app/bibliographic-data/components/priority/priority.enum.ts
export enum Priority {
  /** Международная заявка */
  InternationalApplication = '86',
  /** Публикация международной заявки */
  PublicationInternationalApplication = '87',
  /** Евразийская заявка */
  EurasianApplication = '96',
  /** Публикация евразийской заявки */
  PublicationEurasianApplication = '97',
  /** Приоритет следующей предшествующей заявки (национальная заявка) */
  PriorityNextPendingNationalApplication = '006.1',
  /** Приоритет следующей предшествующей заявки (региональная заявка) */
  PriorityNextPendingRegionalApplication = '006.2',
  /** Приоритет следующей предшествующей заявки (международная заявка) */
  PriorityNextPendingInternationalApplication = '006.3',
  /** Первой(ых) заявки(ок) в государстве-участнике Парижской конвенции (п. 2 ст. 20 Закона) */
  FirstApplicationParisConvention = '20.2',
  /** Поступления дополнительных материалов к более ранней заявки в соответствии с п. 3 ст. 20 Закона */
  EqualsParagraph3Article20 = '20.3',
  /** Более ранней заявки в РГП "НИИС" в соответствии с п. 4 ст. 20 Закона */
  EqualsParagraph4Article20 = '20.4',
  /** Первоначальной заявки в РГП "НИИС" в соответствии с п. 5 ст. 20 Закона */
  EqualsParagraph5Article20 = '20.5',
  /** В соответствии с п. 6 ст. 20 Закона */
  EqualsParagraph6Article20 = '20.6',
  /** Приоритетные данные */
  PriorityData = '30 - 300',
  /** Выставочный приоритет */
  ExhibitionPriority = '45',
  /** Заявляемое обозначение имеет более раннюю регистрацию в Казахстане */
  ClaimedDesignationHasEarlierRegistration = '46',
  /** Поступление дополнительных материалов (правило 6 (2) Инструкции) */
  Rule6_2 = '6.2',
  /** Подача предшествующей евразийской заявки того же заявителя (правило 6 (3) Инструкции) */
  Rule6_3 = '6.3',
  /** Начала открытого показа объекта, содержащего указанное выше изобретение на официальной или официально признанной международной выставке (правило 6 (5) Инструкции) */
  Rule6_5 = '6.5',
  /** Заявка или свидетельство по отношению к которому настоящий документ является дополнительным */
  ApplicationOrCertificateRelationDocumentAdditional = '61',
  /** Более ранняя заявка, из которой выделен настоящий документ */
  EarlierApplicationDocumentAllocated = '62',
  /** Более ранняя заявка, по отношению к которой настоящий патентный документ является продолжением */
  EarlierApplicationPatentDocumentContinuation = '63',
  /** Опубликованный патентный документ, касающийся данной заявки */
  PublishedPatentDocumentRelatingApplication = '65',
  /** Более ранняя заявка, заменой которой является настоящий документ, то есть, если более поздняя заявка подана после отказа по более ранней заявке на то же самое изобретение */
  EarlierApplicationReplacementPresentDocument = '66',
  /** Поступления заявки на допуск к использованию в госкомиссию (п. 2 ст. 4 Закона) */
  ReceiptApplicationAdmissionStateCommission = '4.2',
  /** Подачи первой заявки в стране-участнице Международной конвенции (п. 2 ст. 7 Закона) */
  FilingApplicationCountryInternationalConvention = '7.2',
  /** Предлагался ли сорт (порода) к продаже или продавался в РК */
  WhetherOfferedRepublic = '7.3',
  /** Предлагался ли сорт (порода) к продаже или продавался в других странах */
  WhetherOfferedOther = '7.4',
  /** Международная (Евразийская) Заявка */
  InternationalEurasianApplication = '80'
}