// Iserv.Niis.Angular/src/app/bibliographic-data/components/priority/convention-type.enum.ts
export enum ConventionType {
   /**
   * Нет приоритета
   */
  NoPriority = '02',
  /**
   * Конвенционная
   */
  Convention = '01',
  /**
   * PCT
   */
  PCT = '03',
  /**
   * PCT (Конвенционная)
   */
  PCTConvention = '04',
  /**
   * PCT (Получающее ведомство)
   */
  PCTReceivingOffice = '05',
  /**
   * На основе евразийской заявки
   */
  EurasianApplication = '06'
}