// Iserv.Niis.Angular/src/app/shared/enums/protection-doc-types.enum.ts
export enum ProtectionDocTypes {
  /**
   * Тоговый знак
   */
  Trademark = 'TM',
  /**
   * Изобретение
   */
  Invention = 'B',
  /**
   * Промышленный образец
   */
  IndustrialDesign = 'S2',
  /**
   * Полезная модель
   */
  UsefulModel = 'U',
  /**
   * Селекционное достижение
   */
  SelectionAchievement = 'SA',
  /**
   * Наименование мест происхождения товаров (НМПТ)
   */
  AppellationOfOrigin = 'NPMT',
  /**
   * Международный торговый знак (МТЗ)
   */
  InternationalTrademark = 'TMI'
}